requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.5
paramiko==2.12.0
in_place==0.5.0